refactor: inline openEditor and spawn logic
