+++ packages/cli/src/handlers/ai.ts
@@ -1,3 +1,4 @@
+import { spawn } from "node:child_process";
 import { parseArgs } from "node:util";
 import { createContext, validateWorktreeExists } from "@aku11i/phantom-core";
 import { getGitRoot } from "@aku11i/phantom-git";
@@ -5,16 +6,35 @@ import { getPhantomEnv } from "@aku11i/phantom-process";
 import { isErr } from "@aku11i/phantom-shared";
 import { exitCodes, exitWithError } from "../errors.ts";
 import { output } from "../output.ts";
-import { openEditor } from "../utils/open-editor.ts";
 
 export async function launchAiAssistant(
   aiCommand: string,
   worktreeName: string,
   worktreePath: string,
 ): Promise<number> {
-  return openEditor(aiCommand, [], worktreePath, {
-    ...process.env,
-    ...getPhantomEnv(worktreeName, worktreePath),
+  return new Promise((resolve, reject) => {
+    const child = spawn(aiCommand, [], {
+      cwd: worktreePath,
+      env: {
+        ...process.env,
+        ...getPhantomEnv(worktreeName, worktreePath),
+      },
+      stdio: "inherit",
+      shell: true,
+    });
+
+    child.on("error", (error) => {
+      reject(error);
+    });
+
+    child.on("exit", (code, signal) => {
+      if (signal) {
+        reject(new Error(`Command exited with signal ${signal}`));
+        return;
+      }
+
+      resolve(code ?? 0);
+    });
   });
 }
