--- packages/cli/src/utils/open-editor.ts
@@ -1,31 +1,0 @@
-import { spawn } from "node:child_process";
-
-export async function openEditor(
-  command: string,
-  args: string[],
-  cwd: string,
-  env: NodeJS.ProcessEnv,
-): Promise<number> {
-  return new Promise((resolve, reject) => {
-    const child = spawn(command, args, {
-      // shell:true keeps commands with flags (e.g., "code --wait") working.
-      cwd,
-      env,
-      stdio: "inherit",
-      shell: true,
-    });
-
-    child.on("error", (error) => {
-      reject(error);
-    });
-
-    child.on("exit", (code, signal) => {
-      if (signal) {
-        reject(new Error(`Command exited with signal ${signal}`));
-        return;
-      }
-
-      resolve(code ?? 0);
-    });
-  });
-}
